# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DICOM Gateway API
  description: |
    REST API for DICOM Gateway - A lightweight, HIPAA-compliant Linux DICOM Gateway
    for receiving and forwarding DICOM studies with byte-preserving integrity.
    
    ## Features
    - C-STORE SCP (receiver) and SCU (forwarder) with byte-preserving transmission
    - PostgreSQL-backed job queue with LISTEN/NOTIFY
    - Role-based access control (RBAC)
    - Comprehensive audit logging
    - Prometheus metrics export
    - Configuration management via YAML
  version: 0.1.0
  contact:
    name: DICOM Gateway Support
  license:
    name: Proprietary
servers:
  - url: http://localhost:8000/api/v1
    description: Local development server
  - url: https://gateway.example.com/api/v1
    description: Production server
...
paths:
...
  /config:
    get:
      summary: Get configuration
      description: Get current configuration (sensitive fields redacted)
      tags:
        - Config
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Configuration
          content:
            application/json:
              schema:
                type: object
                properties:
                  config:
                    type: object
                  config_path:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /config/reload:
    post:
      summary: Reload configuration
      description: Reload configuration from file
      tags:
        - Config
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Configuration reloaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigReloadResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /config/upload:
    post:
      summary: Upload configuration file
      description: Upload a new configuration file (YAML)
      tags:
        - Config
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Configuration uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  config_path:
                    type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '400':
          description: Invalid file format
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    ConfigReloadResponse:
      type: object
      properties:
        message:
          type: string
        config_path:
          type: string
...
    Error:
      type: object
      properties:
        error:
          type: string
        detail:
          type: string
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...